+++ src/main/java/com/cloudogu/scm/ssh/command/ScmCommand.java
@@ -24,6 +24,9 @@
 package com.cloudogu.scm.ssh.command;
 
 import com.cloudogu.scm.ssh.Attributes;
+import com.cloudogu.scm.ssh.simplecommand.SimpleCommand;
+import com.cloudogu.scm.ssh.simplecommand.SimpleCommandContext;
+import com.cloudogu.scm.ssh.simplecommand.SimpleCommandFactory;
 import org.apache.shiro.util.ThreadContext;
 import org.apache.sshd.common.channel.ChannelOutputStream;
 import org.apache.sshd.common.util.threads.CloseableExecutorService;
@@ -35,6 +38,7 @@ import sonia.scm.protocolcommand.CommandInterpreter;
 import sonia.scm.protocolcommand.CommandInterpreterFactory;
 import sonia.scm.protocolcommand.RepositoryContext;
 
+import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Optional;
 import java.util.Set;
@@ -44,10 +48,17 @@ public class ScmCommand extends AbstractCommandSupport {
   private static final Logger LOG = LoggerFactory.getLogger(ScmCommand.class);
 
   private final Set<CommandInterpreterFactory> commandInterpreterFactories;
+  private final Set<SimpleCommandFactory> simpleCommandFactories;
 
-  ScmCommand(String command, CloseableExecutorService executorService, Set<CommandInterpreterFactory> commandInterpreterFactories) {
+  ScmCommand(
+    String command,
+    CloseableExecutorService executorService,
+    Set<CommandInterpreterFactory> commandInterpreterFactories,
+    Set<SimpleCommandFactory> simpleCommandFactories
+  ) {
     super(command, executorService);
     this.commandInterpreterFactories = commandInterpreterFactories;
+    this.simpleCommandFactories = simpleCommandFactories;
   }
 
   @Override
@@ -71,25 +82,24 @@ public class ScmCommand extends AbstractCommandSupport {
     String command = getCommand();
     LOG.debug("trying to run command '{}'", command);
     try {
+      int exitCode = 0;
       ThreadContext.bind(getSession().getAttribute(Attributes.SUBJECT));
 
-      CommandInterpreter commandInterpreter = commandInterpreterFactories.stream()
-        .map(p -> p.canHandle(command))
-        .filter(Optional::isPresent)
-        .map(Optional::get)
-        .findFirst()
-        .orElseThrow(() -> new IllegalArgumentException("no interpreter found for command: " + command));
+      Optional<SimpleCommand> optionalSimpleCommand = getSimpleCommand(command);
+      if (optionalSimpleCommand.isPresent()) {
+        exitCode = executeSimpleCommand(optionalSimpleCommand.get(), command);
+      }
 
-      String[] args = commandInterpreter.getParsedArgs();
+      Optional<CommandInterpreter> optionalCommandInterpreter = getCommandInterpreter(command);
+      if (optionalCommandInterpreter.isPresent()) {
+        executeProtocolCommand(optionalCommandInterpreter.get(), command);
+      }
 
-      RepositoryContext repositoryContext = commandInterpreter.getRepositoryContextResolver().resolve(args);
-      CommandContext commandContext = createCommandContext(command, args);
+      if (!optionalSimpleCommand.isPresent() && !optionalCommandInterpreter.isPresent()) {
+        throw new IllegalArgumentException("no interpreter found for command: " + command);
+      }
 
-      commandInterpreter.getProtocolHandler().handle(commandContext, repositoryContext);
-
-      LOG.debug("finished protocol handling of command '{}'", command);
-
-      onExit(0);
+      onExit(exitCode);
     } catch (Exception throwable) {
       LOG.error("received error during handling of command {}", command, throwable);
 
@@ -99,9 +109,35 @@ public class ScmCommand extends AbstractCommandSupport {
     }
   }
 
-  private CommandContext createCommandContext(String command, String[] args) {
-    return new CommandContext(command, args, in, out, err);
+  private int executeSimpleCommand(SimpleCommand simpleCommand, String command) throws IOException {
+    int exitCode = simpleCommand.execute(new SimpleCommandContext(command, in, out, err));
+    LOG.debug("finished execution of command '{}'", command);
+    return exitCode;
   }
 
+  private void executeProtocolCommand(CommandInterpreter commandInterpreter, String command) throws IOException {
+    String[] args = commandInterpreter.getParsedArgs();
+    RepositoryContext repositoryContext = commandInterpreter.getRepositoryContextResolver().resolve(args);
+    CommandContext commandContext = new CommandContext(command, args, in, out, err);
+    commandInterpreter.getProtocolHandler().handle(commandContext, repositoryContext);
+    LOG.debug("finished protocol handling of command '{}'", command);
+  }
 
+  private Optional<SimpleCommand> getSimpleCommand(String command) {
+    return simpleCommandFactories
+      .stream()
+      .map(p -> p.canHandle(command))
+      .filter(Optional::isPresent)
+      .map(Optional::get)
+      .findFirst();
+  }
+
+  private Optional<CommandInterpreter> getCommandInterpreter(String command) {
+    return commandInterpreterFactories
+      .stream()
+      .map(p -> p.canHandle(command))
+      .filter(Optional::isPresent)
+      .map(Optional::get)
+      .findFirst();
+  }
 }