+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add access token command ([#11](https://github.com/scm-manager/scm-ssh-plugin/pull/11))
+
 ## 2.1.0 - 2020-08-20
 ### Added
 - Documentation in English and German ([#6](https://github.com/scm-manager/scm-ssh-plugin/pull/6))